requirements.txt
# Reproducible ML pipeline – frozen versions
# Python 3.11

scikit-learn==1.5.2
pandas==2.2.2
numpy==1.26.4
PyYAML==6.0.2
joblib==1.4.2
